% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrfit.R
\name{qr_fit}
\alias{qr_fit}
\title{The Fitting function of conditional quantile regression model}
\usage{
qr_fit(y, X, tau)
}
\arguments{
\item{y}{The (n * 1) dependent response variable vector of the regression model.}

\item{X}{The (n * p) independent cause variable matrix of the regression model.}

\item{tau}{The constant level chosen according to which quantile needs to be estimated.}
}
\value{
The value of \code{beta_hat} that minimize the objective function of conditional quantile regression model.
}
\description{
Solve for \code{beta_hat} of the minimization problem of the objective function of conditional quantile regression model.
}
