% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrloss.R
\name{qr_loss}
\alias{qr_loss}
\title{The loss function of conditional quantile regression model}
\usage{
qr_loss(z, tau)
}
\arguments{
\item{z}{The value of the objective function needed to be passed to the loss function.}

\item{tau}{The constant level chosen according to which quantile needs to be estimated.}
}
\value{
The value of the loss function with indicator function  I{z < 0}.
}
\description{
Calculates \verb{z*(tau-1\{z<0\})}.
}
