% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrfit-package.R
\docType{package}
\name{qrfit}
\alias{qrfit}
\title{Inference for the Conditional Quantile Regression Model}
\description{
Inference for the Conditional Quantile Regression Model
}
\details{
Provide the Evaluating Function of the Objective Function with Loss Function of the Conditional Quantile Regression model
and the Fitting Function regard to the minimization problem of Conditional Quantile Regression Model.
where the Fitting Function use Linear Programming Algorithm with lpSolve.
}
\keyword{"_PACKAGE"}
\keyword{internal}
